+++ src/components/index.js
@@ -1,1 +1,2 @@
+export { FindConsortiumMember } from './FindConsortiumMember';
 export { SwitchActiveAffiliation } from './SwitchActiveAffiliation';