+++ src/routes/ConsortiumManager.js
@@ -1,3 +1,0 @@
-export const ConsortiumManager = () => {
-  return 'Consortium Manager';
-};