+++ test/jest/fixtures/index.js
@@ -1,2 +1,3 @@
+export * from './affiliations';
 export * from './consortium';
 export * from './tenants';