+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 * [UICONSET-125](https://issues.folio.org/browse/UICONSET-123) Add `ui-consortia-settings` app icon.
 * [UICONSET-7](https://issues.folio.org/browse/UICONSET-7) Create "Consortium manager" interface.
 * [UICONSET-8](https://issues.folio.org/browse/UICONSET-8) Limit access to "Consortium manager" interface.
+* [UICONSET-10](https://issues.folio.org/browse/UICONSET-10) Select member institutions for consortia management.